+++ src/main/java/acme/features/customer/booking/CustomerBookingsController.java
@@ -1,5 +1,5 @@
 
-package acme.features.authenticated.customer.booking;
+package acme.features.customer.booking;
 
 import javax.annotation.PostConstruct;
