+++ src/main/java/acme/features/customer/booking/CustomerBookingsCreateService.java
@@ -1,22 +1,25 @@
 
-package acme.features.authenticated.customer.booking;
+package acme.features.customer.booking;
 
 import java.util.Collection;
+import java.util.Date;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
 import acme.client.components.models.Dataset;
 import acme.client.components.views.SelectChoices;
+import acme.client.helpers.MomentHelper;
 import acme.client.helpers.PrincipalHelper;
 import acme.client.services.AbstractGuiService;
+import acme.client.services.GuiService;
 import acme.entities.booking.Booking;
 import acme.entities.booking.TravelClass;
 import acme.entities.flight.Flight;
 import acme.entities.flight.FlightRepository;
-import acme.entities.passenger.Passenger;
 import acme.realms.Customer;
 
-public class CustomerBookingsPublishService extends AbstractGuiService<Customer, Booking> {
+@GuiService
+public class CustomerBookingsCreateService extends AbstractGuiService<Customer, Booking> {
 
 	// Internal state ---------------------------------------------------------
 
@@ -26,48 +29,57 @@ public class CustomerBookingsPublishService extends AbstractGuiService<Customer,
 	@Autowired
 	private FlightRepository			flightRepository;
 
-	// AbstractGuiService interface -------------------------------------------
+	// AbstractService interface ----------------------------------------------
 
 
 	@Override
 	public void authorise() {
-		int customerId = super.getRequest().getPrincipal().getActiveRealm().getUserAccount().getId();
-		int bookingId = super.getRequest().getData("id", int.class);
-		Booking booking = this.repository.findBookingById(bookingId);
-
-		boolean status = booking.getCustomer().getUserAccount().getId() == customerId && super.getRequest().getPrincipal().hasRealmOfType(Customer.class);
+		boolean status;
+		status = super.getRequest().getPrincipal().hasRealmOfType(Customer.class);
 		super.getResponse().setAuthorised(status);
 	}
 
 	@Override
 	public void load() {
 		Booking booking;
-		int bookingId = super.getRequest().getData("id", int.class);
+		int customerId;
+		Customer customer;
+		Date today = MomentHelper.getCurrentMoment();
 
-		booking = this.repository.findBookingById(bookingId);
+		customerId = super.getRequest().getPrincipal().getActiveRealm().getId();
+		customer = this.repository.findCustomerById(customerId);
+
+		booking = new Booking();
+		booking.setCustomer(customer);
+		booking.setPurchaseMoment(today);
+		booking.setDraftMode(true);
 
 		super.getBuffer().addData(booking);
 	}
 
 	@Override
-	public void bind(final Booking object) {
-		super.bindObject(object, "locatorCode", "purchaseMoment", "travelClass", "lastNibble", "price", "passengers");
+	public void bind(final Booking booking) {
+		assert booking != null;
+		Flight flight;
+		int flightId;
+
+		flightId = super.getRequest().getData("flight", int.class);
+		flight = this.flightRepository.findFlightById(flightId);
+
+		super.bindObject(booking, "locatorCode", "purchaseMoment", "travelClass", "draftMode", "lastNibble", "price");
+		booking.setFlight(flight);
+
 	}
 
 	@Override
 	public void validate(final Booking booking) {
-		if (booking.getLastNibble() == null || booking.getLastNibble().isBlank() || booking.getLastNibble().isEmpty()) {
-			String lastNibbleStored = this.repository.findBookingById(booking.getId()).getLastNibble();
-			if (lastNibbleStored == null || lastNibbleStored.isBlank() || lastNibbleStored.isBlank())
-				super.state(false, lastNibbleStored, "acme.validation.confirmation.message.lastNibble");
-		}
+		Booking b = this.repository.findBookingByLocatorCode(booking.getLocatorCode());
+		if (b != null)
+			super.state(false, "locatorCode", "acme.validation.confirmation.message.booking.locator-code");
 	}
 
 	@Override
 	public void perform(final Booking booking) {
-		if (booking.getLastNibble() == null || booking.getLastNibble().isBlank() || booking.getLastNibble().isEmpty())
-			booking.setLastNibble(this.repository.findBookingById(booking.getId()).getLastNibble());
-		booking.setDraftMode(false);
 		this.repository.save(booking);
 	}
 
@@ -80,12 +92,9 @@ public class CustomerBookingsPublishService extends AbstractGuiService<Customer,
 		Collection<Flight> flights = this.flightRepository.findAllFlights();
 		flightChoices = SelectChoices.from(flights, "descriptions", booking.getFlight());
 		choices = SelectChoices.from(TravelClass.class, booking.getTravelClass());
-		Collection<Passenger> passengerNumber = this.repository.findPassengersByBookingId(booking.getId());
-		Collection<String> passengers = passengerNumber.stream().map(x -> x.getFullName()).toList();
 
 		dataset = super.unbindObject(booking, "locatorCode", "purchaseMoment", "draftMode", "lastNibble", "price");
 		dataset.put("travelClass", choices);
-		dataset.put("passengers", passengers);
 		dataset.put("flight", flightChoices.getSelected().getKey());
 		dataset.put("flights", flightChoices);
