+++ src/main/java/acme/features/customer/booking/CustomerBookingsListService.java
@@ -1,5 +1,5 @@
 
-package acme.features.authenticated.customer.booking;
+package acme.features.customer.booking;
 
 import java.util.Collection;
