+++ src/main/java/acme/features/customer/booking/CustomerBookingsUpdateService.java
@@ -1,5 +1,5 @@
 
-package acme.features.authenticated.customer.booking;
+package acme.features.customer.booking;
 
 import java.util.Collection;
 
@@ -17,7 +17,7 @@ import acme.entities.passenger.Passenger;
 import acme.realms.Customer;
 
 @GuiService
-public class CustomerBookingsShowService extends AbstractGuiService<Customer, Booking> {
+public class CustomerBookingsUpdateService extends AbstractGuiService<Customer, Booking> {
 
 	// Internal state ---------------------------------------------------------
 
@@ -51,6 +51,25 @@ public class CustomerBookingsShowService extends AbstractGuiService<Customer, Bo
 	}
 
 	@Override
+	public void bind(final Booking object) {
+		super.bindObject(object, "locatorCode", "purchaseMoment", "draftMode", "lastNibble", "price", "travelClass", "flight");
+	}
+
+	@Override
+	public void validate(final Booking booking) {
+		if (booking.isDraftMode() == false)
+			super.state(false, "draftMode", "acme.validation.confirmation.message.update");
+		Booking b = this.repository.findBookingByLocatorCode(booking.getLocatorCode());
+		if (b != null)
+			super.state(false, "locatorCode", "acme.validation.confirmation.message.booking.locator-code");
+	}
+
+	@Override
+	public void perform(final Booking booking) {
+		this.repository.save(booking);
+	}
+
+	@Override
 	public void unbind(final Booking booking) {
 		Dataset dataset;
 		SelectChoices choices;
@@ -70,5 +89,4 @@ public class CustomerBookingsShowService extends AbstractGuiService<Customer, Bo
 
 		super.getResponse().addData(dataset);
 	}
-
 }
