+++ src/main/java/acme/features/customer/bookingRecord/CustomerBookingRecordController.java
@@ -1,5 +1,5 @@
 
-package acme.features.authenticated.customer.bookingRecord;
+package acme.features.customer.bookingRecord;
 
 import javax.annotation.PostConstruct;
 
@@ -7,11 +7,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import acme.client.controllers.AbstractGuiController;
 import acme.client.controllers.GuiController;
-import acme.entities.booking.Booking;
+import acme.entities.booking.BookingRecord;
 import acme.realms.Customer;
 
 @GuiController
-public class CustomerBookingRecordController extends AbstractGuiController<Customer, Booking> {
+public class CustomerBookingRecordController extends AbstractGuiController<Customer, BookingRecord> {
 
 	// Internal state ---------------------------------------------------------
