+++ src/main/resources/WEB-INF/data/sample/booking-record.csv
@@ -1,0 +1,30 @@
+key,key:booking,key:passenger
+booking-record-01,booking-01,passenger-01
+booking-record-02,booking-02,passenger-02
+booking-record-03,booking-03,passenger-03
+booking-record-04,booking-04,passenger-04
+booking-record-05,booking-05,passenger-05
+booking-record-06,booking-06,passenger-06
+booking-record-07,booking-07,passenger-07
+booking-record-08,booking-08,passenger-08
+booking-record-09,booking-09,passenger-09
+booking-record-10,booking-10,passenger-10
+booking-record-11,booking-11,passenger-11
+booking-record-12,booking-12,passenger-12
+booking-record-13,booking-13,passenger-13
+booking-record-14,booking-14,passenger-14
+booking-record-15,booking-15,passenger-15
+booking-record-16,booking-16,passenger-16
+booking-record-17,booking-17,passenger-17
+booking-record-18,booking-18,passenger-18
+booking-record-19,booking-19,passenger-19
+booking-record-20,booking-20,passenger-20
+booking-record-21,booking-21,passenger-21
+booking-record-22,booking-22,passenger-22
+booking-record-23,booking-23,passenger-23
+booking-record-24,booking-24,passenger-24
+booking-record-25,booking-25,passenger-25
+booking-record-26,booking-26,passenger-26
+bookingRecord-27,booking-27,passenger-27
+bookingRecord-28,booking-28,passenger-28
+bookingRecord-29,booking-29,passenger-29
