+++ src/main/resources/WEB-INF/views/customer/booking/messages-en.i18n
@@ -1,0 +1,32 @@
+# messages-en.i18n
+#
+# Copyright (C) 2012-2025 Rafael Corchuelo.
+#
+# In keeping with the traditional purpose of furthering education and research, it is
+# the policy of the copyright owner to permit non-commercial use and redistribution of
+# this software. It has been tested carefully, but it is not guaranteed for any particular
+# purposes.  The copyright owner does not offer any warranties or representations, nor do
+# they accept any liabilities with respect to them.
+
+customer.booking.list.title = List of bookings
+customer.booking.form.title = List of bookings
+
+booking.list.label.locatorCode = locator Code
+booking.list.label.purchaseMoment = Purchase Moment
+booking.list.label.travelClass = Travel Class
+booking.list.label.lastNibble = Last Nibble
+booking.list.label.price = Price
+booking.list.label.passenger = Passengers
+booking.list.label.draftMode = draftMode
+booking.list.label.flight = flight
+
+booking.form.button.create = Create booking
+booking.form.button.update = Update booking
+booking.form.button.publish = Publish booking
+booking.form.button.passenger = List of passengers
+booking-record.form.button.create = Create a booking record
+# booking.form.button.delete = Delete booking
+
+acme.validation.confirmation.message.lastNibble = The lastNibble must be filled if the booking is published
+acme.validation.confirmation.message.update = A booking can not be updated if it is published
+acme.validation.confirmation.message.booking.locator-code = There is already another booking with this locator code
