+++ src/main/resources/WEB-INF/views/customer/booking/messages-es.i18n
@@ -1,0 +1,32 @@
+# messages-es.i18n
+#
+# Copyright (C) 2012-2025 Rafael Corchuelo.
+#
+# In keeping with the traditional purpose of furthering education and research, it is
+# the policy of the copyright owner to permit non-commercial use and redistribution of
+# this software. It has been tested carefully, but it is not guaranteed for any particular
+# purposes.  The copyright owner does not offer any warranties or representations, nor do
+# they accept any liabilities with respect to them.
+
+customer.booking.list.title = Listado de Reservas
+customer.booking.form.title = Detalles de la Reserva
+
+booking.list.label.locatorCode = Código de localización
+booking.list.label.purchaseMoment = Momento de compra
+booking.list.label.travelClass = Tipo de viaje
+booking.list.label.lastNibble = Ultima tarjeta
+booking.list.label.price = Precio
+booking.list.label.passenger = Pasajeros
+booking.list.label.draftMode = Estado de publicación
+booking.list.label.flight = vuelo
+
+booking.form.button.create = Crear reserva
+booking.form.button.update = Actualizar reserva
+booking.form.button.publish = Publicar reserva
+booking.form.button.passenger = Lista de pasajeros
+booking-record.form.button.create = Crear un registro de reserva
+booking.form.button.delete = Eliminar reserva
+
+acme.validation.confirmation.message.lastNibble = Los últimos digitos de la tarjeta de crédito deberán ser rellenados si la reserva es publicada
+acme.validation.confirmation.message.update = Una reserva no puede ser actualizada si está publicada
+acme.validation.confirmation.message.booking.locator-code = Ya existe otra reserva con este código de localizador
