+++ src/main/resources/WEB-INF/views/customer/bookingRecord/messages-en.i18n
@@ -1,0 +1,17 @@
+# messages-en.i18n
+#
+# Copyright (C) 2012-2025 Rafael Corchuelo.
+#
+# In keeping with the traditional purpose of furthering education and research, it is
+# the policy of the copyright owner to permit non-commercial use and redistribution of
+# this software. It has been tested carefully, but it is not guaranteed for any particular
+# purposes.  The copyright owner does not offer any warranties or representations, nor do
+# they accept any liabilities with respect to them.
+
+customer.booking-record.list.title = List of bookings record
+customer.booking-record.form.title = Create a booking record
+
+customer.booking-record.list.label.passenger = Passenger
+customer.booking-record.list.label.booking = Booking
+
+acme.validation.confirmation.message.booking-record.create = That relation is already created
